Some tests successfull after major refactoring.
